Adding container registry secret API
- API will add secret for remote repo when User and PW are set.

When values are present, the api is invoked automatically when creating
an NS, and the secret is installed in all NS.

When values are present, the App definition templates will also be hinted
to add ImagePullSecrets with the name of the secret (which is constant).

Suggested use:
ie.
export CTR_REGISTRY=usracr.azurecr.io
export CTR_REGISTRY_USER=usracr
export CTR_REGISTRY_PASSWORD=y67V3vB48156qIGaXTomf0fJT6J/zfJv