Catalog2 skeleton
This adds a new Catalog2. It brings the dispatcher from original Catalog; other than that
just ensures that all API calls to SMISpec, Catalog and Controller are first issued through
Catalog2, and then just passhtrough'd directly on the original implementations of the original
modules.

Introduces a few constructs for the shim layer to be effective even through internal
API invocations inside the original modules.

The most notable functional change is the top level RWLock introduced between ADS
and configuration threads, which will be needed to avoid concurrent map accesses to
the data model later on.

I deliberately chose RWLock because it elegantly provides the synchronization needed
when we have single writer / multiple readers (1 dispatcher(config updated), 1-routine per proxy(ADS readers)), and
is also implemented as write-preferred which will prevent write starvation from ADS readers.
Other locking implementations exist, and perf implications are TBD.

Currently the lock is of no real use because there is no data model. Check the commit on top
which implements IsMonitoredNamespace in the data model.