Add POC trafficTarget Namespace mapping
This commit introduces a mechanism to keep track relations between
namespaces, relating them through TrafficTargets.

It allows deriving a subset of affected namespaces based on any config change,
and using these relations an additional subset of proxies can be looked up and
issued for update when a config change is seen in the related namespace.

This is a POC to show that deriving deltas and computing subset of envoys
to be updated at a time is possible, and the level of complexity involved is
probably proportional to the level of detail in terms of being able to
identify affected envoys. Mapping Traffic targets to namespaces is probably
one of the simplest ways of doing it; of course has a number of shortcoming.

At best, if the deployed topologies do not connect all namespaces in a single mesh
through traffic targets (say, deploys individual server and clients that talk in groups
of namespaces), the benefit is the greatest and most notable, and ADS times remain
fairly constant regardless of the number of pods in the mesh.
More than 1000 pods/proxies, and pretty effortlessly at that, with plenty of cpu to
spare, has been reached through this patch despite a number of other issues.

At worst, if all namespaces are fully meshed, or everything is deployed in a single
namespace, this commit shows no improvement as all pods in a namspace are still going
to be touched.

For POC only.